fix: 데이터 타입 변경, update 안 되는 문제 해결

+++ AutoBid_BE/src/main/java/com/codesquad/autobid/car/domain/CheckCarResponse.java
@@ -8,7 +8,7 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor(access = AccessLevel.PRIVATE)
 public class CheckCarResponse {
 
-    private Long distance;
+    private Float distance;
     private String name;
     private Type type;
     private String sellName;
